+++ tests/run/test_run.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import contextlib
-import datetime
 import io
 import pathlib
 import sys
@@ -23,7 +22,7 @@ import pytest
 import mlrun
 import mlrun.errors
 import mlrun.launcher.factory
-from mlrun import MLClientCtx, new_function, new_task
+from mlrun import new_function, new_task
 from tests.conftest import (
     examples_path,
     has_secrets,
@@ -248,158 +247,6 @@ async def test_local_args(db, db_session):
     verify_state(result)
 
     assert output.find(", --xyz, 789") != -1, "params not detected in argv"
-
-
-def test_local_context(rundb_mock):
-    project_name = "xtst"
-    mlrun.mlconf.artifact_path = out_path
-    context = mlrun.get_or_create_ctx("xx", project=project_name, upload_artifacts=True)
-    db = mlrun.get_run_db()
-    run = db.read_run(context._uid, project=project_name)
-    assert run["struct"]["status"]["state"] == "running", "run status not updated in db"
-
-    with context:
-        context.log_artifact("xx", body="123", local_path="a.txt")
-        context.log_model("mdl", body="456", model_file="mdl.pkl", artifact_path="+/mm")
-        context.get_param("p1", 1)
-        context.get_param("p2", "a string")
-        context.log_result("accuracy", 16)
-        context.set_label("label-key", "label-value")
-        context.set_annotation("annotation-key", "annotation-value")
-        context._set_input("input-key", "input-url")
-
-        artifact = context.get_cached_artifact("xx")
-        artifact.format = "z"
-        context.update_artifact(artifact)
-
-    assert context._state == "completed", "task did not complete"
-
-    run = db.read_run(context._uid, project=project_name)
-    run = run["struct"]
-
-    # run state should not be updated by the context
-    assert run["status"]["state"] == "running", "run status was updated in db"
-    assert (
-        run["status"]["artifacts"][0]["metadata"]["key"] == "xx"
-    ), "artifact not updated in db"
-    assert (
-        run["status"]["artifacts"][0]["spec"]["format"] == "z"
-    ), "run/artifact attribute not updated in db"
-    assert run["status"]["artifacts"][1]["spec"]["target_path"].startswith(
-        out_path
-    ), "artifact not uploaded to subpath"
-
-    db_artifact = db.read_artifact(artifact.db_key, project=project_name)
-    assert db_artifact["spec"]["format"] == "z", "artifact attribute not updated in db"
-
-    assert run["spec"]["parameters"]["p1"] == 1, "param not updated in db"
-    assert run["spec"]["parameters"]["p2"] == "a string", "param not updated in db"
-    assert run["status"]["results"]["accuracy"] == 16, "result not updated in db"
-    assert run["metadata"]["labels"]["label-key"] == "label-value", "label not updated"
-    assert (
-        run["metadata"]["annotations"]["annotation-key"] == "annotation-value"
-    ), "annotation not updated"
-
-    assert run["spec"]["inputs"]["input-key"] == "input-url", "input not updated"
-
-
-def test_context_from_dict_when_start_time_is_string():
-    context = mlrun.get_or_create_ctx("ctx")
-    context_dict = context.to_dict()
-    context = mlrun.MLClientCtx.from_dict(context_dict)
-    assert isinstance(context._start_time, datetime.datetime)
-
-
-@pytest.mark.parametrize(
-    "is_api",
-    [True, False],
-)
-def test_context_from_run_dict(is_api):
-    mlrun.config.is_running_as_api = Mock(return_value=is_api)
-    run_dict = {
-        "metadata": {
-            "name": "test-context-from-run-dict",
-            "project": "default",
-            "labels": {"label-key": "label-value"},
-            "annotations": {"annotation-key": "annotation-value"},
-        },
-        "spec": {
-            "parameters": {"p1": 1, "p2": "a string"},
-            "inputs": {
-                "input-key": "input-url",
-                "store-input": "store://store-input",
-            },
-        },
-    }
-    runtime = mlrun.runtimes.base.BaseRuntime.from_dict(run_dict)
-    handler = "my_func"
-    out_path = "test_artifact_path"
-    launcher = mlrun.launcher.factory.LauncherFactory.create_launcher(
-        runtime._is_remote
-    )
-    run = launcher._create_run_object(run_dict)
-    run = launcher._enrich_run(
-        runtime,
-        run,
-        handler,
-        run_dict["metadata"]["project"],
-        run_dict["metadata"]["name"],
-        run_dict["spec"]["parameters"],
-        run_dict["spec"]["inputs"],
-        returns="",
-        hyperparams=None,
-        hyper_param_options=None,
-        verbose=False,
-        scrape_metrics=None,
-        out_path=out_path,
-        artifact_path="",
-        workdir="",
-    )
-    context = MLClientCtx.from_dict(run.to_dict(), is_api=is_api)
-    assert context.name == run_dict["metadata"]["name"]
-    assert context._project == run_dict["metadata"]["project"]
-    assert context._labels == run_dict["metadata"]["labels"]
-    assert context._annotations == run_dict["metadata"]["annotations"]
-    assert context.get_param("p1") == run_dict["spec"]["parameters"]["p1"]
-    assert context.get_param("p2") == run_dict["spec"]["parameters"]["p2"]
-    assert (
-        context.get_input("input-key").artifact_url
-        == run_dict["spec"]["inputs"]["input-key"]
-    )
-    assert context._inputs["input-key"] == run_dict["spec"]["inputs"]["input-key"]
-    # not using get_input because it will try to load the artifact from the nop db
-    assert context._inputs["store-input"] == run_dict["spec"]["inputs"]["store-input"]
-    assert context.labels["label-key"] == run_dict["metadata"]["labels"]["label-key"]
-    assert (
-        context.annotations["annotation-key"]
-        == run_dict["metadata"]["annotations"]["annotation-key"]
-    )
-    assert context.artifact_path == out_path
-
-
-@pytest.mark.parametrize(
-    "state, error, expected_state",
-    [
-        ("running", None, "completed"),
-        ("completed", None, "completed"),
-        (None, "error message", "error"),
-        (None, "", "error"),
-    ],
-)
-def test_context_set_state(rundb_mock, state, error, expected_state):
-    project_name = "test_context_error"
-    mlrun.mlconf.artifact_path = out_path
-    context = mlrun.get_or_create_ctx("xx", project=project_name, upload_artifacts=True)
-    db = mlrun.get_run_db()
-    run = db.read_run(context._uid, project=project_name)
-    assert run["struct"]["status"]["state"] == "running", "run status not updated in db"
-
-    with context:
-        context.set_state(execution_state=state, error=error, commit=False)
-        context.commit(completed=True)
-
-    assert context._state == expected_state, "task state was not set correctly"
-    assert context._error == error, "task error was not set"
 
 
 def test_run_class_code():
